Add a method to set flags on Pagure PRs
Flags on PRs are specific to Pagure.

Other forges allow setting flags/statuses only on commits and will
display the flags for the PR head on the PR page, if any is set.

Due to the above this method is added only to PagurePullRequests, and we
expect OGR library users to implement the "show the flags of the last
commit for the PR" behaviour for Pagure, in case they need it.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>