GitHub: query only one user when telling if a user can merge PRs
When telling if a user can merge PRs in a GitHub repo, ask GitHub for
the user's permission on the repo instead of checking if the user is in
the list of collaborators.

This should lead to fewer API calls in cases when the full list of
collaborators is not needed, but it also returns better results: the
list of collaborators does not seem to be always up to date when queries
are done using GitHub App tokens.

Relates to packit-service/packit-service#612.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>