Introduce indirect decorator
In many cases we have specialized classes that have static methods providing
functionality that we query from GitProject. In majority of cases, it is
just a copy-paste of parameters to the call of specialized static method.

Introducing an indirect decorator that takes specialized static method
which takes the project (or parent object) as first argument and all other
parameters.

Greatly reduces unnecessary boilerplate for such methods and keeps the
specific code contained in its own class.

Signed-off-by: Matej Focko <mfocko@redhat.com>