Convert common errors to ogr exceptions
Some exceptions, such as authentication errors can occur in a lot of
places (e.g. when a token expires). Trying to catch it everywhere would
be tedious. Catch these exceptions and network errors in a common
metaclass and convert them to an ogr exception thrown to the user to
avoid code duplication

Signed-off-by: František Nečas <fnecas@redhat.com>