Create Bodhi updates for branched development version as well
Rawhide updates are created automatically,
but not for not-released branched versions.

We can use `fedora-latest` alias that is the latest non-rawhide branch.
It is not an issue if `fedora-latest` is covered also by `fedora-stable`,
Packit converts all the values to a set
so we don't need to worry about duplicates.

Signed-off-by: Frantisek Lachman <flachman@redhat.com>