fix(pagure): check codes instead of textual reason
Fixes #947

During the merge of packit/packit-service#2845 I've noticed that the
tests failed with the following contents of logs:

    self = <PagureProject(namespace="rpms", repo="python-teamcity-messages")>
    path = 'python-teamcity-messages.spec', ref = 'HEAD'

        def get_file_content(self, path: str, ref=None) -> str:
            ref = ref or self.default_branch
            result = self._call_project_api_raw(
                "raw",
                ref,
                "f",
                path,
                add_api_endpoint_part=False,
            )

            if not result or result.reason == "NOT FOUND":
                raise FileNotFoundError(f"File '{path}' on {ref} not found")
            if result.reason != "OK":
    >           raise PagureAPIException(
                    f"File '{path}' on {ref} not found due to {result.reason}",
                )
    E           ogr.exceptions.PagureAPIException: File 'python-teamcity-messages.spec' on HEAD not found due to Not Found

As it appears, textual response is apparently not case-sensitive, which
is also supported by RFC 9110 (section 15.1.):

    The status codes listed below are defined in this specification.
    The reason phrases listed here are only recommendations -- they can
    be replaced by local equivalents or left out altogether without
    affecting the protocol.

Therefore it's a better idea to check `.ok` field provided by `requests`,
or directly against the status code (which is actually what's been caught
by accident in the Service tests).

Since I'm already touching this issue, I'm also converting the checks of
`.reason == "OK"` to `.ok` that is present in the responses too.

---

RFC 9110: https://datatracker.ietf.org/doc/html/rfc9110.html#section-15.1-1

Signed-off-by: Matej Focko <mfocko@packit.dev>